refactor(core): background_task 迁出为 task 插件，收敛 #208 Phase 1d
- 新建 crates/plugins/tiangong-plugin-task（plugin.rs + handler.rs + lib.rs）
- handler.rs: 搬原 tool/background_task.rs 的 TaskRegistry/TaskStatus/TaskInfo/task_registry/wait_tasks
  task_registry() 保持全局 OnceLock 静态，GUI 管理与 LLM 工具命中同一注册表
- plugin.rs: TaskPlugin 实现 Plugin + ToolSpecProvider + ToolOverrideHandler
  - tool_specs 提供 spawn_task/query_task/list_tasks/cancel_task/wait_tasks 5 个 spec
  - handle dispatch 搬自 runtime.rs::handle_background_task
  - register 通过 engine.runtime_env_handle() 拿 env 共享句柄（照 command plugin 模式）
- 不再需要 BackgroundTaskHandler trait：ToolOverrideHandler 已是那个抽象，
  registry 自动把 tool_specs 返回的工具名绑定到 handle
- core 清理 3 处硬编码特判：
  - runtime.rs:415 删 handle_background_task 短路调用
  - runtime.rs:452-693 删 handle_background_task 整个函数（~250 行）
  - runtime.rs:717-750 删 inject_enhanced_tools 的 5 个后台任务 spec
- core 删 src/tool/background_task.rs + tool.rs 的 pub mod 声明
- context/assembler.rs 的 3 个过滤函数留 core（纯字符串匹配，按工具名过滤）
- src-tauri/commands.rs: task_registry() 改引 tiangong_plugin_task
- 三入口（src-tauri/app.rs、cli/repl.rs、server/remote/core.rs）注册 task plugin
