+++ crates/plugins/tiangong-plugin-task/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "tiangong-plugin-task"
+version.workspace = true
+edition = "2024"
+license = "Apache-2.0"
+
+[dependencies]
+tiangong-core = { workspace = true }
+tiangong-types = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+scru128 = { workspace = true }
+tracing = { workspace = true }
