+++ crates/tiangong-cli/Cargo.toml
@@ -17,6 +17,7 @@ tiangong-plugin-text-to-speech = { workspace = true }
 tiangong-plugin-speech-to-text = { workspace = true }
 tiangong-plugin-memory = { workspace = true }
 tiangong-plugin-scheduler = { workspace = true }
+tiangong-plugin-task = { workspace = true }
 tiangong-plugin-analyze-attachment = { workspace = true }
 tiangong-plugin-skill = { workspace = true }
 tiangong-plugin-mcp = { workspace = true }
