+++ crates/tiangong-core/src/runtime.rs
@@ -411,10 +411,8 @@ impl RuntimeEngine {
             }
         }
 
-        // 后台任务管理
-        if let Some(result) = self.handle_background_task(call) {
-            return result;
-        }
+        // 后台任务管理（spawn_task / query_task / list_tasks / cancel_task / wait_tasks）
+        // 已迁移至独立插件 crate（tiangong-plugin-task），由 tool_overrides 统一分发。
 
         // Skill 详情查询（get_skill_detail）已迁移至独立插件 crate
         // （tiangong-plugin-skill），由 tool_overrides 统一分发。
@@ -448,252 +446,6 @@ impl RuntimeEngine {
         }
     }
 
-    /// 处理后台任务工具调用
-    fn handle_background_task(&self, call: &ToolCall) -> Option<ToolResult> {
-        use crate::tool::background_task::{TaskStatus, task_registry};
-
-        match call.name.as_str() {
-            "spawn_task" => {
-                let name = call
-                    .arguments
-                    .get("name")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("task")
-                    .to_string();
-                let cmd = call
-                    .arguments
-                    .get("cmd")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or_default()
-                    .to_string();
-                let args: Vec<String> = call
-                    .arguments
-                    .get("args")
-                    .and_then(|v| v.as_array())
-                    .map(|arr| {
-                        arr.iter()
-                            .filter_map(|v| v.as_str())
-                            .map(String::from)
-                            .collect()
-                    })
-                    .unwrap_or_default();
-                let cwd = call
-                    .arguments
-                    .get("cwd")
-                    .and_then(|v| v.as_str())
-                    .map(String::from);
-
-                if cmd.is_empty() {
-                    return Some(ToolResult {
-                        ok: false,
-                        summary: "缺少 cmd 参数".to_string(),
-                        stdout: String::new(),
-                        stderr: String::new(),
-                        exit_code: 1,
-                        execution: None,
-                    });
-                }
-
-                let env = self
-                    .runtime_env
-                    .lock()
-                    .map(|g| {
-                        g.iter()
-                            .map(|(k, v)| (k.clone(), v.clone()))
-                            .collect::<Vec<_>>()
-                    })
-                    .unwrap_or_default();
-
-                match task_registry().lock() {
-                    Ok(mut reg) => match reg.spawn(name, cmd, args, cwd, env) {
-                        Ok(task_id) => Some(ToolResult {
-                            ok: true,
-                            summary: format!("后台任务已启动，task_id={task_id}"),
-                            stdout: serde_json::json!({"task_id": task_id}).to_string(),
-                            stderr: String::new(),
-                            exit_code: 0,
-                            execution: None,
-                        }),
-                        Err(e) => Some(ToolResult {
-                            ok: false,
-                            summary: format!("启动后台任务失败：{e}"),
-                            stdout: String::new(),
-                            stderr: e,
-                            exit_code: 1,
-                            execution: None,
-                        }),
-                    },
-                    Err(e) => Some(ToolResult {
-                        ok: false,
-                        summary: format!("任务注册表锁失败：{e}"),
-                        stdout: String::new(),
-                        stderr: e.to_string(),
-                        exit_code: 1,
-                        execution: None,
-                    }),
-                }
-            }
-            "query_task" => {
-                let task_id = call
-                    .arguments
-                    .get("task_id")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or_default();
-                match task_registry().lock() {
-                    Ok(mut reg) => match reg.query(task_id) {
-                        Some(info) => {
-                            let status_text = match &info.status {
-                                TaskStatus::Running => "running".to_string(),
-                                TaskStatus::Completed { exit_code } => {
-                                    format!("completed (exit_code={exit_code})")
-                                }
-                                TaskStatus::Failed { error } => format!("failed: {error}"),
-                                TaskStatus::Cancelled => "cancelled".to_string(),
-                            };
-                            Some(ToolResult {
-                                ok: true,
-                                summary: format!("任务 {} 状态：{}", info.name, status_text),
-                                stdout: serde_json::to_string_pretty(&info).unwrap_or_default(),
-                                stderr: String::new(),
-                                exit_code: 0,
-                                execution: None,
-                            })
-                        }
-                        None => Some(ToolResult {
-                            ok: false,
-                            summary: format!("未找到任务：{task_id}"),
-                            stdout: String::new(),
-                            stderr: String::new(),
-                            exit_code: 1,
-                            execution: None,
-                        }),
-                    },
-                    Err(e) => Some(ToolResult {
-                        ok: false,
-                        summary: e.to_string(),
-                        stdout: String::new(),
-                        stderr: String::new(),
-                        exit_code: 1,
-                        execution: None,
-                    }),
-                }
-            }
-            "list_tasks" => match task_registry().lock() {
-                Ok(mut reg) => {
-                    let tasks = reg.list();
-                    Some(ToolResult {
-                        ok: true,
-                        summary: format!("{} 个后台任务", tasks.len()),
-                        stdout: serde_json::to_string_pretty(&tasks).unwrap_or_default(),
-                        stderr: String::new(),
-                        exit_code: 0,
-                        execution: None,
-                    })
-                }
-                Err(e) => Some(ToolResult {
-                    ok: false,
-                    summary: e.to_string(),
-                    stdout: String::new(),
-                    stderr: String::new(),
-                    exit_code: 1,
-                    execution: None,
-                }),
-            },
-            "cancel_task" => {
-                let task_id = call
-                    .arguments
-                    .get("task_id")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or_default();
-                match task_registry().lock() {
-                    Ok(mut reg) => match reg.cancel(task_id) {
-                        Some(info) => Some(ToolResult {
-                            ok: true,
-                            summary: format!("任务 {} 已取消", info.name),
-                            stdout: serde_json::to_string_pretty(&info).unwrap_or_default(),
-                            stderr: String::new(),
-                            exit_code: 0,
-                            execution: None,
-                        }),
-                        None => Some(ToolResult {
-                            ok: false,
-                            summary: format!("未找到任务：{task_id}"),
-                            stdout: String::new(),
-                            stderr: String::new(),
-                            exit_code: 1,
-                            execution: None,
-                        }),
-                    },
-                    Err(e) => Some(ToolResult {
-                        ok: false,
-                        summary: e.to_string(),
-                        stdout: String::new(),
-                        stderr: String::new(),
-                        exit_code: 1,
-                        execution: None,
-                    }),
-                }
-            }
-            "wait_tasks" => {
-                let task_ids: Vec<String> = call
-                    .arguments
-                    .get("task_ids")
-                    .and_then(|v| v.as_array())
-                    .map(|arr| {
-                        arr.iter()
-                            .filter_map(|v| v.as_str())
-                            .map(String::from)
-                            .collect()
-                    })
-                    .unwrap_or_default();
-                let timeout_ms = call
-                    .arguments
-                    .get("timeout_ms")
-                    .and_then(|v| v.as_u64())
-                    .unwrap_or(0);
-
-                if task_ids.is_empty() {
-                    return Some(ToolResult {
-                        ok: false,
-                        summary: "缺少 task_ids 参数".to_string(),
-                        stdout: String::new(),
-                        stderr: String::new(),
-                        exit_code: 1,
-                        execution: None,
-                    });
-                }
-
-                let results = crate::tool::background_task::wait_tasks(task_ids, timeout_ms);
-                let all_ok = results.iter().all(
-                    |r| matches!(r.status, TaskStatus::Completed { exit_code } if exit_code == 0),
-                );
-                let running_count = results
-                    .iter()
-                    .filter(|r| matches!(r.status, TaskStatus::Running))
-                    .count();
-                let summary = if running_count > 0 {
-                    format!(
-                        "{} 个任务完成，{} 个仍在运行（超时）",
-                        results.len() - running_count,
-                        running_count
-                    )
-                } else {
-                    format!("{} 个任务全部完成", results.len())
-                };
-
-                Some(ToolResult {
-                    ok: all_ok,
-                    summary,
-                    stdout: serde_json::to_string_pretty(&results).unwrap_or_default(),
-                    stderr: String::new(),
-                    exit_code: if all_ok { 0 } else { 1 },
-                    execution: None,
-                })
-            }
-            _ => None, // 不是后台任务工具
-        }
-    }
-
     pub fn fallback_error_message(err: &anyhow::Error) -> String {
         format!("执行失败：{err}")
     }
@@ -714,40 +466,9 @@ pub(crate) fn inject_enhanced_tools(tools: &mut Vec<ToolSpec>) {
     // install_skill / remove_skill / set_skill_enabled 不再作为 LLM 工具暴露——它们
     // 是 app 层（Tauri 命令 / CLI modal / app_state facade）使用的管理 API。
 
-    // 后台任务管理
-    for spec in [
-        (
-            "spawn_task",
-            "在后台启动特殊命令。仅当用户明确要求后台、不阻塞、并行执行、持续运行、启动服务/监听，或需要让命令跨多轮继续运行时使用；普通命令、构建、检查、git、文件操作必须优先使用 run_shell 或 run_command。",
-            serde_json::json!({"type":"object","properties":{"name":{"type":"string"},"cmd":{"type":"string"},"args":{"type":"array","items":{"type":"string"}},"cwd":{"type":"string"}},"required":["name","cmd"]}),
-        ),
-        (
-            "query_task",
-            "查询后台任务状态",
-            serde_json::json!({"type":"object","properties":{"task_id":{"type":"string"}},"required":["task_id"]}),
-        ),
-        (
-            "list_tasks",
-            "列出所有后台任务",
-            serde_json::json!({"type":"object","properties":{}}),
-        ),
-        (
-            "cancel_task",
-            "取消后台任务",
-            serde_json::json!({"type":"object","properties":{"task_id":{"type":"string"}},"required":["task_id"]}),
-        ),
-        (
-            "wait_tasks",
-            "等待已通过 spawn_task 启动的后台任务完成。仅用于已有后台任务，不用于执行普通命令。",
-            serde_json::json!({"type":"object","properties":{"task_ids":{"type":"array","items":{"type":"string"}},"timeout_ms":{"type":"integer"}},"required":["task_ids"]}),
-        ),
-    ] {
-        tools.push(ToolSpec {
-            name: spec.0.to_string(),
-            description: spec.1.to_string(),
-            input_schema: spec.2,
-        });
-    }
+    // 后台任务管理（spawn_task / query_task / list_tasks / cancel_task / wait_tasks）
+    // 已迁移至独立插件 crate（tiangong-plugin-task），由 ToolSpecProvider 提供 spec。
+
     // MCP 管理（register/remove/set_mcp_enabled）已迁移至独立插件 crate
     // （tiangong-plugin-mcp），是 app 层管理 API，不再作为 LLM 工具暴露。
