+++ crates/tiangong-core/src/tool.rs
@@ -4,12 +4,11 @@
 //! - list_dir / read_file / write_file 等 8 个 → tiangong-plugin-fs
 //! - web_fetch → tiangong-plugin-fetch（CLI/Server）/ browser 插件（GUI）
 //! - run_command / run_shell → tiangong-plugin-command（CLI/Server）/ terminal 插件（GUI）
+//! - spawn_task / query_task / list_tasks / cancel_task / wait_tasks → tiangong-plugin-task
 //!
 //! core 不再直接执行任何工具，仅保留 ToolResult 供插件 handler 返回。
-//! background_task（spawn_task 后台任务）仍由 core 的 runtime.rs 直接处理，
 //! common（路径沙箱/命令白名单）暴露给插件 crate 复用。
 
-pub mod background_task;
 pub mod common;
 pub(crate) mod media;
 pub use common::{session_workspace_root, set_session_cwd};
